# openapi/api.yaml
openapi: 3.0.3
info:
  title: LLM Gateway API
  description: |
    A unified API gateway for interacting with multiple LLM providers (OpenAI, Anthropic, etc.)
    with built-in compliance checking, rate limiting, session management, and audit logging.
  version: 1.0.0
  contact:
    name: LLM Gateway Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.example.com/api/v1
    description: Production server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api-keys:
    post:
      tags:
        - api-keys
      summary: Create a new API key
      description: |
        Create a new API key bound to a specific team, developer, and release unit.
        The developer must be pre-authorized for the release unit.
        The raw API key is only returned once - store it securely.
      operationId: createAPIKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAPIKeyRequest'
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAPIKeyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: Developer not authorized for release unit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - api-keys
      summary: List API keys
      description: List all API keys for the authenticated developer.
      operationId: listAPIKeys
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/APIKeyInfo'
  /api-keys/{id}:
    delete:
      tags:
        - api-keys
      summary: Revoke an API key
      description: Revoke an API key by ID. The key will no longer be usable.
      operationId: revokeAPIKey
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: API key revoked
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /release-units/authorize:
    post:
      tags:
        - release-units
      summary: Authorize developer for release units
      description: |
        Grant a developer permission to create API keys for one or more release units.
        This is typically done by team admins. Accepts an array of release unit names.
      operationId: authorizeReleaseUnit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizeReleaseUnitRequest'
      responses:
        '201':
          description: Authorization granted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "release unit authorized"
        '400':
          $ref: '#/components/responses/BadRequest'
  /release-units/revoke:
    delete:
      tags:
        - release-units
      summary: Revoke release unit authorization
      description: Revoke a developer's authorization for a release unit.
      operationId: revokeReleaseUnit
      parameters:
        - name: team_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: developer_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: release_unit
          in: query
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Authorization revoked
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'API key for authentication (format: Bearer {api_key})'
...
  schemas:
...
    CreateAPIKeyRequest:
      type: object
      required:
        - team_id
        - developer_id
        - user_id
        - release_unit
      properties:
        team_id:
          type: string
          format: uuid
        developer_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        release_unit:
          type: string
          description: Project/release unit name
          example: "project-alpha"
        description:
          type: string
          description: Optional description for the key
          example: "CI/CD pipeline key"
        expires_in_days:
          type: integer
          description: Days until expiration (optional)
          example: 90
    CreateAPIKeyResponse:
      type: object
      properties:
        api_key:
          type: string
          description: The raw API key - only shown once, store securely
          example: "llmgw_a1b2c3d4e5f6..."
        key_prefix:
          type: string
          description: First 8 chars for identification
          example: "llmgw_a1"
        team_id:
          type: string
          format: uuid
        developer_id:
          type: string
          format: uuid
        release_unit:
          type: string
        expires_at:
          type: string
          format: date-time
    APIKeyInfo:
      type: object
      properties:
        id:
          type: integer
        key_prefix:
          type: string
          description: First 8 chars for identification
        team_id:
          type: string
          format: uuid
        developer_id:
          type: string
          format: uuid
        release_unit:
          type: string
        description:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
    AuthorizeReleaseUnitRequest:
      type: object
      required:
        - team_id
        - developer_id
        - release_units
      properties:
        team_id:
          type: string
          format: uuid
        developer_id:
          type: string
          format: uuid
        release_units:
          type: array
          items:
            type: string
          minItems: 1
          description: List of project/release unit names to authorize
          example: ["project-alpha", "project-beta", "project-gamma"]
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "invalid request body"
...